+++ lib/src/metta/runner/arithmetics.rs
@@ -308,33 +308,8 @@ macro_rules! def_binary_bool_op {
 def_binary_bool_op!(AndOp, and, &&);
 def_binary_bool_op!(OrOp, or, ||);
 
-// NOTE: xor and flip are absent in Python intentionally for conversion testing
+// NOTE: xor is absent in Python intentionally for conversion testing
 def_binary_bool_op!(XorOp, xor, ^);
-
-#[derive(Clone, PartialEq, Debug)]
-pub struct FlipOp{}
-
-impl Display for FlipOp {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "flip")
-    }
-}
-
-impl Grounded for FlipOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, ATOM_TYPE_BOOL])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for FlipOp {
-    fn execute(&self, _args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        Ok(vec![Atom::gnd(Bool(rand::random()))])
-    }
-}
 
 
 #[derive(Clone, PartialEq, Debug)]
